Show/hide modals when navigating with back/forward buttons
Using pushState and popState event handlers, we can understand when the
modal needs to be closed (and prevents the browser from navigating
away), and when the modal needs to be reopened (using the new modal:reopen
event).

By default this is turned off, but can be enabled in the modal options
with `updateHistory: true`.

Note that as this stands, this doesn't work very well with Turbolinks:
I'm not sure yet if this is a bug on the Turbolinks side or not, but in
the meantime, this behaviour works for non-Turbolinks sites.